added user validation, permissions api

+++ app/api/users/index.js
@@ -12,6 +12,10 @@ router.get('/current', function(req, res) {
 });
 
 router.get('/users', function(req, res) {
+  if (!req.currentUser.isAdmin) {
+    return res.status(403).send('You don\'t have permissions');
+  }
+
   var page = req.query.page;
   var search = req.query.search;
 
@@ -26,6 +30,10 @@ router.get('/users', function(req, res) {
 });
 
 router.get('/user/:id', function(req, res) {
+  if (!req.currentUser.isAdmin) {
+    return res.status(403).send('You don\'t have permissions');
+  }
+
   var id = req.params.id;
 
   editorManager.getUserEditor({ id: id })
@@ -38,6 +46,10 @@ router.get('/user/:id', function(req, res) {
 });
 
 router.post('/user', function(req, res) {
+  if (!req.currentUser.isAdmin) {
+    return res.status(403).send('You don\'t have permissions');
+  }
+
   var user = req.body.user;
   usersManager.save({
     user: user
@@ -45,12 +57,19 @@ router.post('/user', function(req, res) {
   .then(function(user) {
     return res.json({ user: user });
   })
-  .catch(function() {
+  .catch(function(error) {
+    if (error) {
+      return res.status(500).send(error);
+    }
     return res.status(500).send('Can\'t save');
   });
 });
 
 router.patch('/user/:id', function(req, res) {
+  if (!req.currentUser.isAdmin) {
+    return res.status(403).send('You don\'t have permissions');
+  }
+
   var id = req.params.id;
   var user = req.body.user;
 
@@ -61,12 +80,19 @@ router.patch('/user/:id', function(req, res) {
   .then(function(user) {
     return res.json({ user: user });
   })
-  .catch(function() {
-    return res.status(500).send('Can\'t patch');
+  .catch(function(error) {
+    if (error) {
+      return res.status(500).send(error);
+    }
+    return res.status(500).send('Can\'t save');
   });
 });
 
 router.delete('/user/:id', function(req, res) {
+  if (!req.currentUser.isAdmin) {
+    return res.status(403).send('You don\'t have permissions');
+  }
+
   var id = req.params.id;
 
   usersManager.deleteUser({ id: id })
